converge join logic, crate a draft of options and rewriter
