+++ parquet-hadoop/src/main/java/org/apache/parquet/hadoop/rewrite/RewriteOptions.java
@@ -18,10 +18,7 @@
  */
 package org.apache.parquet.hadoop.rewrite;
 
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 import java.util.stream.Collectors;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.Path;
@@ -39,11 +36,20 @@ import org.apache.parquet.io.OutputFile;
 
 /**
  * A set of options to create a ParquetRewriter.
+ *
+ * TODO find a place where to put a proper description of functionality as it is not trivial:
+ * ParquetRewriter allows to stitch files with the same schema into a single file.
+ * Note that ParquetRewriter also can be used for effectively stitch/joining multiple parquet files with
+ * different schemas.
+ * You can provide the main input file group and multiple right side ones. That is possible when:
+ * 1) the number of rows in the main and extra input groups are the same,
+ * 2) the ordering of rows in the main and extra input groups is the same.
  */
 public class RewriteOptions {
 
   private final ParquetConfiguration conf;
   private final List<InputFile> inputFiles;
+  private final List<List<InputFile>> inputFilesR;
   private final OutputFile outputFile;
   private final List<String> pruneColumns;
   private final CompressionCodecName newCodecName;
@@ -55,6 +61,7 @@ public class RewriteOptions {
   private RewriteOptions(
       ParquetConfiguration conf,
       List<InputFile> inputFiles,
+      List<List<InputFile>> inputFilesR,
       OutputFile outputFile,
       List<String> pruneColumns,
       CompressionCodecName newCodecName,
@@ -64,6 +71,7 @@ public class RewriteOptions {
       IndexCache.CacheStrategy indexCacheStrategy) {
     this.conf = conf;
     this.inputFiles = inputFiles;
+    this.inputFilesR = inputFilesR;
     this.outputFile = outputFile;
     this.pruneColumns = pruneColumns;
     this.newCodecName = newCodecName;
@@ -110,6 +118,27 @@ public class RewriteOptions {
         .collect(Collectors.toList());
   }
 
+  /** TODO fix documentation after addition of inputFilesR
+   * Gets the right input {@link Path}s for the rewrite if they exist for all input files,
+   * otherwise throws a {@link RuntimeException}.
+   *
+   * @return a {@link List} of the associated right input {@link Path}s
+   */
+  public List<List<Path>> getInputFilesR() {
+    return inputFilesR.stream()
+        .map(x -> x.stream()
+          .map(y -> {
+            if (y instanceof HadoopOutputFile) {
+              HadoopOutputFile hadoopOutputFile = (HadoopOutputFile) y;
+              return new Path(hadoopOutputFile.getPath());
+            } else {
+              throw new RuntimeException("The input files do not all have an associated Hadoop Path.");
+            }
+          }).collect(Collectors.toList())
+        )
+        .collect(Collectors.toList());
+  }
+
   /**
    * Gets the {@link InputFile}s for the rewrite.
    *
@@ -117,6 +146,16 @@ public class RewriteOptions {
    */
   public List<InputFile> getParquetInputFiles() {
     return inputFiles;
+  }
+
+
+  /** TODO fix documentation after addition of inputFilesR
+   * Gets the right {@link InputFile}s for the rewrite.
+   *
+   * @return a {@link List} of the associated right {@link InputFile}s
+   */
+  public List<List<InputFile>> getParquetInputFilesR() {
+    return inputFilesR;
   }
 
   /**
@@ -170,6 +209,7 @@ public class RewriteOptions {
   public static class Builder {
     private final ParquetConfiguration conf;
     private final List<InputFile> inputFiles;
+    private final List<List<InputFile>> inputFilesR = new ArrayList<>();
     private final OutputFile outputFile;
     private List<String> pruneColumns;
     private CompressionCodecName newCodecName;
@@ -325,6 +365,21 @@ public class RewriteOptions {
       return this;
     }
 
+    /** TODO fix documentation after addition of inputFilesR
+     * Add an input file to read from.
+     *
+     * @param paths input file path to read from
+     * @return self
+     */
+    public Builder addInputPathsR(List<Path> paths) {
+      this.inputFilesR.add(
+          paths.stream()
+              .map(x -> HadoopInputFile.fromPathUnchecked(x, ConfigurationUtil.createHadoopConfiguration(conf)))
+              .collect(Collectors.toList())
+      );
+      return this;
+    }
+
     /**
      * Add an input file to read from.
      *
@@ -333,6 +388,17 @@ public class RewriteOptions {
      */
     public Builder addInputFile(InputFile inputFile) {
       this.inputFiles.add(inputFile);
+      return this;
+    }
+
+    /** TODO fix documentation after addition of inputFilesR
+     * Add an input file to read from.
+     *
+     * @param inputFiles input file to read from
+     * @return self
+     */
+    public Builder addInputFilesR(List<InputFile> inputFiles) {
+      this.inputFilesR.add(inputFiles);
       return this;
     }
 
@@ -358,6 +424,16 @@ public class RewriteOptions {
     public RewriteOptions build() {
       Preconditions.checkArgument(inputFiles != null && !inputFiles.isEmpty(), "Input file is required");
       Preconditions.checkArgument(outputFile != null, "Output file is required");
+      Preconditions.checkArgument(inputFilesR.stream().allMatch(x -> x != null && !x.isEmpty()),
+          "Right side Input files can't be empty, if you don't need a join functionality then use other builders");
+      Preconditions.checkArgument(inputFilesR.isEmpty() || pruneColumns == null,
+          "Right side Input files join functionality does not yet support column pruning");
+      Preconditions.checkArgument(inputFilesR.isEmpty() || maskColumns == null,
+          "Right side Input files join functionality does not yet support column masking");
+      Preconditions.checkArgument(inputFilesR.isEmpty() || encryptColumns == null,
+          "Right side Input files join functionality does not yet support column encryption");
+      Preconditions.checkArgument(inputFilesR.isEmpty() || newCodecName == null,
+          "Right side Input files join functionality does not yet support codec changing");
 
       if (pruneColumns != null) {
         if (maskColumns != null) {
@@ -390,6 +466,7 @@ public class RewriteOptions {
       return new RewriteOptions(
           conf,
           inputFiles,
+          inputFilesR,
           outputFile,
           pruneColumns,
           newCodecName,
